// mims-web/core/modules/setting/model/traffic/infrastructure/ModelTrafficParameterRequest.d.ts
export interface ITrafficParams {
	directional_distribution_factor: number
	elane: number
	four_wheel_axle_number: number
	four_wheel_vehicle_volume: number
	is_truck_factor: boolean
	lane_distribution_factor: number
	road_group_id: number
	six_wheel_axle_number_id: number
	six_wheel_factor_result: number
	six_wheel_percentage_truck: number
	six_wheel_vehicle_volume: number
	speed_average: number
	speed_heavy_truck: number
	ten_wheel_axle_number_id: number
	ten_wheel_factor_result: number
	ten_wheel_percentage_truck: number
	ten_wheel_vehicle_volume: number
}